+++ src/vue-router/redirect.js
@@ -1,16 +1,20 @@
 import { vueRoutes } from './routes'
 
 const REDIRECT_TYPES = {
-  email: 1,
+  apiVerify: 'api-verify',
+  inviteVerify: 'invite-verify',
 }
 
 export const resolveRedirect = (to, from, next) => {
   const encodedValue = to.path.replace('/r/', '')
-  const action = JSON.parse(atob(encodedValue))
+  const decodedValue = JSON.parse(atob(encodedValue))
 
-  switch (action.type) {
-    case REDIRECT_TYPES.email:
+  switch (decodedValue.type) {
+    case REDIRECT_TYPES.apiVerify:
       handleEmailRedirect(encodedValue, next)
+      break
+    case REDIRECT_TYPES.inviteVerify:
+      handleInviteRedirect(decodedValue.meta, next)
       break
   }
 }
@@ -19,5 +23,12 @@ function handleEmailRedirect (encodedVerificationCode, next) {
   next({
     ...vueRoutes.login,
     params: { encodedVerificationCode },
+  })
+}
+
+function handleInviteRedirect (decodedInviteVerificationInfo, next) {
+  next({
+    ...vueRoutes.signup,
+    params: { inviteVerificationInfo: decodedInviteVerificationInfo },
   })
 }