+++ src/vue/forms/WithdrawalForm.vue
@@ -24,11 +24,16 @@
                 </option>
               </select-field>
               <div class="withdrawal__form-field-description">
-                <p>
+                <p
+                  :class="{
+                    'withdrawal__form-field-description--error'
+                      :!balance
+                  }"
+                >
                   {{
                     'withdrawal-form.balance' | globalize({
                       balance: {
-                        value: selectedAssetBalance.balance,
+                        value: balance,
                         currency: selectedAssetBalance.asset.code
                       }
                     })
@@ -54,7 +59,7 @@
                 asset: form.asset.code
               })"
               :asset="form.asset"
-              :disabled="formMixin.isDisabled"
+              :disabled="formMixin.isDisabled || !balance"
               is-max-button-shown
             />
           </div>
@@ -72,7 +77,7 @@
                 :label="'withdrawal-form.destination-address' | globalize({
                   asset: form.asset.code
                 })"
-                :disabled="formMixin.isDisabled"
+                :disabled="formMixin.isDisabled || !balance"
               />
             </template>
 
@@ -85,7 +90,7 @@
                 @blur="touchField('form.address')"
                 :error-message="getFieldErrorMessage('form.address')"
                 :label="'withdrawal-form.comment' | globalize"
-                :disabled="formMixin.isDisabled"
+                :disabled="formMixin.isDisabled || !balance"
               />
             </template>
           </div>
@@ -106,7 +111,7 @@
               v-if="!formMixin.isConfirmationShown"
               type="submit"
               class="app__button-raised"
-              :disabled="formMixin.isDisabled"
+              :disabled="formMixin.isDisabled || !balance"
             >
               {{ 'withdrawal-form.withdraw-btn' | globalize }}
             </button>
@@ -217,7 +222,9 @@ export default {
       withdrawableBalancesAssets: vuexTypes.withdrawableBalancesAssets,
       accountBalanceByCode: vuexTypes.accountBalanceByCode,
     }),
-
+    balance () {
+      return +this.selectedAssetBalance.balance
+    },
     isMasterAssetOwner () {
       return this.form.asset.owner === api.networkDetails.adminAccountId
     },
@@ -345,6 +352,10 @@ export default {
 .withdrawal__form-field-description {
   margin-top: 0.7rem;
   opacity: 0.7;
+}
+
+.withdrawal__form-field-description--error {
+  color: $col-error;
 }
 
 .withdrawal__fee-table {