+++ CHANGELOG.md
@@ -15,6 +15,7 @@ for further information about branching and tagging conventions.
 - redirect types
 - number input field
 - Bug with max value of amount input field for deposit
+- A bug with wrong displaying asset code in requests asset creation table
 
 ## [1.12.0-rc.1] - 2020-03-25
 #### Added