+++ CHANGELOG.md
@@ -10,6 +10,7 @@ for further information about branching and tagging conventions.
 #### Added
 - Redirect type invite-verify
 - Transfer, withdraw and issuance forms disabling if balance is 0
+- Validation to registration, sign in, recovery, change password input fields
 
 #### Fixed
 - redirect types
@@ -18,6 +19,7 @@ for further information about branching and tagging conventions.
 - Visibility of select field options on top-bar if width < 768px
 - Bug with displaying correct unlocked amount in movements
 - Text align in security tab in settings for mobile
+- A bug with env config
 
 ## [1.12.0-rc.1] - 2020-03-25
 #### Added